Move date functions and tests to helpers
