+++ src/Components/Details/Details.component.tsx
@@ -1,19 +1,12 @@
 import React from "react";
 
-import {
-    format,
-    isSameWeek,
-    isSameDay,
-    isSameHour,
-    isSameMinute,
-} from "date-fns";
-
 import { Button } from "@zendeskgarden/react-buttons";
 import { Grid, Row, Col } from "@zendeskgarden/react-grid";
 import { Well, Title, Paragraph } from "@zendeskgarden/react-notifications";
 import { Tag } from "@zendeskgarden/react-tags";
 import { PencilIcon, XIcon } from "@heroicons/react/solid";
 
+import { formatDateTime } from "@helper/date";
 import { Work } from "@model/work";
 
 type Props = {
@@ -67,27 +60,6 @@ const Details: React.FC<Props> = (props: Props) => {
             </Grid>
         </Well>
     );
-};
-
-export const formatDateTime = (d: Date, dur?: number): string => {
-    return dur && dur > 0 ?
-        `${formatRelativeDate(d)} for ${dur} minutes.` :
-        `${formatRelativeDate(d)}.`;
-};
-
-export const formatRelativeDate = (d: Date): string => {
-    const now = new Date();
-    if (isSameMinute(d, now)) {
-        return "A few seconds ago";
-    } else if (isSameHour(d, now)) {
-        return `This hour at ${format(d, "HH:mm")}`;
-    } else if (isSameDay(d, now)) {
-        return `Today at ${format(d, "HH:mm")}`;
-    } else if (isSameWeek(d, now)) {
-        return `${format(d, "EEEE")} at ${format(d, "HH:mm")}`;
-    } else {
-        return format(d, "do MMMM yyyy HH:mm");
-    }
 };
 
 export default Details;
